+++ hosts/default/config.nix
@@ -1,20 +1,19 @@
+# 💫 https://github.com/JaKooLit 💫 #
 # Main default config
+
+
+# NOTE!!! : Packages and Fonts are configured in packages-&-fonts.nix
+
 
 { config, pkgs, host, username, options, lib, inputs, system, ...}: let
   
   inherit (import ./variables.nix) keyboardLayout;
-  python-packages = pkgs.python3.withPackages (
-    ps:
-      with ps; [
-        requests
-        pyquery # needed for hyprland-dots Weather script
-        ]
-    );
-  
+    
   in {
   imports = [
     ./hardware.nix
     ./users.nix
+    ./packages-&-fonts.nix
     ../../modules/amd-drivers.nix
     ../../modules/nvidia-drivers.nix
     ../../modules/nvidia-prime-drivers.nix
@@ -188,96 +187,6 @@
   users = {
     mutableUsers = true;
   };
-
-  environment.systemPackages = (with pkgs; [
-  # System Packages
-    baobab
-    btrfs-progs
-    clang
-    curl
-    cpufrequtils
-    duf
-    eza
-    ffmpeg   
-    glib #for gsettings to work
-    gsettings-qt
-    git
-    killall  
-    libappindicator
-    libnotify
-    openssl #required by Rainbow borders
-    pciutils
-    vim
-    wget
-    xdg-user-dirs
-    xdg-utils
-
-    fastfetch
-    (mpv.override {scripts = [mpvScripts.mpris];}) # with tray
-    #ranger
-      
-    # Hyprland Stuff
-    (ags.overrideAttrs (oldAttrs: {
-        inherit (oldAttrs) pname;
-        version = "1.8.2";
-      }))
-    #ags    
-    btop
-    brightnessctl # for brightness control
-    #cava
-    cliphist
-    eog
-    gnome-system-monitor
-    file-roller
-    grim
-    gtk-engine-murrine #for gtk themes
-    hyprcursor # requires unstable channel
-    hypridle # requires unstable channel
-    imagemagick 
-    inxi
-    jq
-    kitty
-    libsForQt5.qtstyleplugin-kvantum #kvantum
-    networkmanagerapplet
-    nwg-look # requires unstable channel
-    nvtopPackages.full
-    pamixer
-    pavucontrol
-    playerctl
-    polkit_gnome
-    pyprland
-    libsForQt5.qt5ct
-    qt6ct
-    qt6.qtwayland
-    qt6Packages.qtstyleplugin-kvantum #kvantum
-    rofi-wayland
-    slurp
-    swappy
-    swaynotificationcenter
-    swww
-    unzip
-    wallust
-    wl-clipboard
-    wlogout
-    yad
-    yt-dlp
-
-    #waybar  # if wanted experimental next line
-    #(pkgs.waybar.overrideAttrs (oldAttrs: { mesonFlags = oldAttrs.mesonFlags ++ [ "-Dexperimental=true" ];}))
-  ]) ++ [
-	  python-packages
-  ];
-
-  # FONTS
-  fonts.packages = with pkgs; [
-    noto-fonts
-    fira-code
-    noto-fonts-cjk-sans
-    jetbrains-mono
-    font-awesome
-	  terminus_font
-    nerd-fonts.jetbrains-mono
- 	];
 
   # Extra Portal Configuration
   xdg.portal = {
@@ -502,5 +411,5 @@
   # this value at the release version of the first install of this system.
   # Before changing this value read the documentation for this option
   # (e.g. man configuration.nix or on https://nixos.org/nixos/options.html).
-  system.stateVersion = "24.05"; # Did you read the comment?
+  system.stateVersion = "24.11"; # Did you read the comment?
 }