+++ hosts/default/packages-&-fonts.nix
@@ -1,0 +1,104 @@
+# 💫 https://github.com/JaKooLit 💫 #
+# Packages and Fonts config
+
+{ pkgs, inputs, ...}: let
+
+  python-packages = pkgs.python3.withPackages (
+    ps:
+      with ps; [
+        requests
+        pyquery # needed for hyprland-dots Weather script
+        ]
+    );
+
+  in {
+  environment.systemPackages = (with pkgs; [
+  # System Packages
+    baobab
+    btrfs-progs
+    clang
+    curl
+    cpufrequtils
+    duf
+    eza
+    ffmpeg   
+    glib #for gsettings to work
+    gsettings-qt
+    git
+    killall  
+    libappindicator
+    libnotify
+    openssl #required by Rainbow borders
+    pciutils
+    vim
+    wget
+    xdg-user-dirs
+    xdg-utils
+
+    fastfetch
+    (mpv.override {scripts = [mpvScripts.mpris];}) # with tray
+    #ranger
+      
+    # Hyprland Stuff
+    (ags.overrideAttrs (oldAttrs: {
+        inherit (oldAttrs) pname;
+        version = "1.8.2";
+      }))
+    #ags    
+    btop
+    brightnessctl # for brightness control
+    #cava
+    cliphist
+    eog
+    gnome-system-monitor
+    file-roller
+    grim
+    gtk-engine-murrine #for gtk themes
+    hyprcursor # requires unstable channel
+    hypridle # requires unstable channel
+    imagemagick 
+    inxi
+    jq
+    kitty
+    libsForQt5.qtstyleplugin-kvantum #kvantum
+    networkmanagerapplet
+    nwg-look # requires unstable channel
+    nvtopPackages.full
+    pamixer
+    pavucontrol
+    playerctl
+    polkit_gnome
+    pyprland
+    libsForQt5.qt5ct
+    qt6ct
+    qt6.qtwayland
+    qt6Packages.qtstyleplugin-kvantum #kvantum
+    rofi-wayland
+    slurp
+    swappy
+    swaynotificationcenter
+    swww
+    unzip
+    wallust
+    wl-clipboard
+    wlogout
+    yad
+    yt-dlp
+
+    #waybar  # if wanted experimental next line
+    #(pkgs.waybar.overrideAttrs (oldAttrs: { mesonFlags = oldAttrs.mesonFlags ++ [ "-Dexperimental=true" ];}))
+  ]) ++ [
+	  python-packages
+  ];
+
+  # FONTS
+  fonts.packages = with pkgs; [
+    noto-fonts
+    fira-code
+    noto-fonts-cjk-sans
+    jetbrains-mono
+    font-awesome
+	  terminus_font
+    nerd-fonts.jetbrains-mono
+ 	];
+  }