+++ hosts/default/users.nix
@@ -1,3 +1,6 @@
+# 💫 https://github.com/JaKooLit 💫 #
+# Users - NOTE: Packages defined on this will be on current user only
+
 { pkgs, username, ... }:
 
 let